+++ ignite/templates/module/migration/templates.go
@@ -1,0 +1,6 @@
+package modulemigration
+
+import "embed"
+
+//go:embed files/* files/**/*
+var files embed.FS